Implémentation des services

+++ src/main/java/fr/upsaclay/easychair/repository/OrganizerRepository.java
@@ -3,5 +3,5 @@ package fr.upsaclay.easychair.repository;
 import fr.upsaclay.easychair.model.Organizer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrganizerRepository extends JpaRepository<Organizer, Integer> {
+public interface OrganizerRepository extends JpaRepository<Organizer, Long> {
 }
